rename certbot to acme-authorizer
Fixes #2
