+++ README.md
@@ -1,5 +1,5 @@
 ## Overview
-This gem is an implementation of the [certbot http-01
+This gem is an implementation of the [ACME http-01
 challenge](https://tools.ietf.org/html/draft-ietf-acme-acme-01#page-38)
 for use in Rails applications.
 
@@ -7,10 +7,10 @@ for use in Rails applications.
 
 ```ruby
 # Gemfile
-gem 'certbot'
+gem 'acme-authorizer'
 ```
 
-NOTE: the certbot challenge/response *must* be served over HTTP without
+NOTE: the ACME http-01 challenge/response *must* be served over HTTP without
 SSL.  This means that your Rails application can not have the `force_ssl` flag
 set in your `config/application.rb`.
 
@@ -19,21 +19,21 @@ http://guides.rubyonrails.org/action_controller_overview.html#force-https-protoc
 
 ## Configuration
 
-By default, this library is configured via pairs of ENV variables with the format:
+By default, this library is configured via pairs of ENV variables with the same format used by [sabayon](https://github.com/dmathieu/sabayon):
 ```
-/CERTBOT_TOKEN_[0-9]+/
-/CERTBOT_KEY_AUTHORIZATION_[0-9]+/
+/ACME_TOKEN_[0-9]+/
+/ACME_KEY_[0-9]+/
 ```
 
 for example:
 ```
-CERTBOT_TOKEN_0=123123
-CERTBOT_KEY_AUTHORIZATION_0=123123
+ACME_TOKEN_0=123123
+ACME_KEY_0=123123
 ```
 
 The challenge and token can also be configured via Ruby API.
 ```ruby
-Certbot.configure do |config|
+Acme::Authorizer.configure do |config|
   config.add_token('my_challenge_token', 'my_key_authorization')
 end
 ```
