+++ app/controllers/acme/authorizer/application_controller.rb
@@ -1,0 +1,7 @@
+module Acme
+  module Authorizer
+    class ApplicationController < ActionController::Base
+      protect_from_forgery with: :exception
+    end
+  end
+end
