--- app/controllers/certbot/application_controller.rb
@@ -1,5 +1,0 @@
-module Certbot
-  class ApplicationController < ActionController::Base
-    protect_from_forgery with: :exception
-  end
-end
