--- app/helpers/certbot/application_helper.rb
@@ -1,4 +1,0 @@
-module Certbot
-  module ApplicationHelper
-  end
-end
