--- app/views/layouts/certbot/application.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Certbot</title>
-  <%= stylesheet_link_tag    "certbot/application", media: "all" %>
-  <%= javascript_include_tag "certbot/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= yield %>
-
-</body>
-</html>
