+++ config/routes.rb
@@ -1,3 +1,3 @@
 Rails.application.routes.draw do
-  get '/.well-known/acme-challenge/:token' => 'certbot#show', constraints: { protocol: 'http://' }
+  get '/.well-known/acme-challenge/:token' => 'acme/authorizer/tokens#show', constraints: { protocol: 'http://' }
 end
