+++ lib/acme/authorizer/version.rb
@@ -1,0 +1,5 @@
+module Acme
+  module Authorizer
+    VERSION = "0.0.1.pre.3"
+  end
+end
