--- lib/certbot/version.rb
@@ -1,3 +1,0 @@
-module Certbot
-  VERSION = "0.0.1.pre.3"
-end
