--- lib/tasks/certbot_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :certbot do
-#   # Task goes here
-# end
