+++ spec/dummy/config/application.rb
@@ -9,7 +9,7 @@ require "sprockets/railtie"
 # require "rails/test_unit/railtie"
 
 Bundler.require(*Rails.groups)
-require "certbot"
+require "acme-authorizer"
 
 module Dummy
   class Application < Rails::Application
