+++ spec/routing/certbot_routing_spec.rb
@@ -1,3 +1,3 @@
-RSpec.describe CertbotController do
-  it { expect(get: '/.well-known/acme-challenge/123123').to route_to('certbot#show', token: '123123', protocol: 'http://') }
+RSpec.describe Acme::Authorizer::TokensController do
+  it { expect(get: '/.well-known/acme-challenge/123123').to route_to('acme/authorizer/tokens#show', token: '123123', protocol: 'http://') }
 end
